Cast as 32 rather than 16

+++ src/delay.c
@@ -19,6 +19,6 @@ void adc_init(){
 }
 
 void update_delay(){
-    playback_delay = MIN_DELAY + (((uint16_t)(MAX_DELAY - MIN_DELAY) * ADC0.RESULT) >> 8);
+    playback_delay = MIN_DELAY + (((uint32_t)(MAX_DELAY - MIN_DELAY) * ADC0.RESULT) >> 8);
     //printf("%d\n", playback_delay);
 }
